Split up state.rs
